remove observation in locomotion task mdp
